--- pkgs/base/swarmauri_base/transports/transport_base.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-from .TransportBase import CapabilityError, TransportBase, _require_caps
-
-__all__ = ["CapabilityError", "TransportBase", "_require_caps"]
