+++ pkgs/core/swarmauri_core/transports/__init__.py
@@ -3,10 +3,13 @@
 from .ITransport import ITransport
 from .enums import AddressScheme, Cast, Feature, IOModel, Protocol, SecurityMode
 from .i_anycast import IAnycastTransport
+from .i_appserver import HttpApp, IAppServer
 from .i_broadcast import IBroadcastTransport
+from .i_http_client import IHttpClient
 from .i_multicast import IMulticastTransport
 from .i_multiplex import IMultiplexTransport, ChannelHandle
 from .i_peer import IPeerTransport
+from .i_runnable import Handler, IRunnable, Role
 from .i_unicast import IUnicastTransport
 from .capabilities import TransportCapabilities
 
@@ -17,13 +20,19 @@ __all__ = [
     "Feature",
     "IAnycastTransport",
     "IBroadcastTransport",
+    "Handler",
+    "HttpApp",
+    "IAppServer",
+    "IHttpClient",
     "IMulticastTransport",
     "IMultiplexTransport",
     "ITransport",
     "IOModel",
+    "IRunnable",
     "IPeerTransport",
     "IUnicastTransport",
     "Protocol",
+    "Role",
     "SecurityMode",
     "TransportCapabilities",
 ]
