+++ pkgs/standards/swarmauri_transport_h2/swarmauri_transport_h2/h2_transport.py
@@ -6,7 +6,7 @@ from typing import Dict, List, Optional, Tuple
 from swarmauri_base.transports.http_client_mixin import HttpClientMixin
 from swarmauri_base.transports.http_server_mixin import HttpServerMixin
 from swarmauri_base.transports.runnable_mixin import RunnableMixin
-from swarmauri_base.transports.transport_base import TransportBase
+from swarmauri_base.transports.TransportBase import TransportBase
 from swarmauri_core.transports.capabilities import TransportCapabilities
 from swarmauri_core.transports.enums import Feature
