+++ pkgs/standards/swarmauri_transport_meshsidecarhttp2/swarmauri_transport_meshsidecarhttp2/mesh_sidecar_h2.py
@@ -6,7 +6,7 @@ from typing import Optional
 
 from swarmauri_base.transports.peer_mixin import PeerTransportMixin
 from swarmauri_base.transports.runnable_mixin import RunnableMixin
-from swarmauri_base.transports.transport_base import TransportBase
+from swarmauri_base.transports.TransportBase import TransportBase
 from swarmauri_core.transports.capabilities import TransportCapabilities
 from swarmauri_core.transports.enums import (
     AddressScheme,
