+++ pkgs/standards/swarmauri_transport_mtlsunicast/swarmauri_transport_mtlsunicast/mtls_unicast.py
@@ -7,7 +7,7 @@ from typing import Optional
 
 from swarmauri_base.transports.peer_mixin import PeerTransportMixin
 from swarmauri_base.transports.runnable_mixin import RunnableMixin
-from swarmauri_base.transports.transport_base import TransportBase
+from swarmauri_base.transports.TransportBase import TransportBase
 from swarmauri_base.transports.unicast_mixin import UnicastTransportMixin
 from swarmauri_core.transports.capabilities import TransportCapabilities
 from swarmauri_core.transports.enums import (
