+++ pkgs/standards/swarmauri_transport_sseoutbound/swarmauri_transport_sseoutbound/sse_outbound.py
@@ -6,7 +6,7 @@ import ssl
 from typing import Optional, Set
 
 from swarmauri_base.transports.runnable_mixin import RunnableMixin
-from swarmauri_base.transports.transport_base import TransportBase
+from swarmauri_base.transports.TransportBase import TransportBase
 from swarmauri_core.transports.capabilities import TransportCapabilities
 from swarmauri_core.transports.enums import (
     AddressScheme,
