+++ pkgs/standards/swarmauri_transport_sshtunnel/swarmauri_transport_sshtunnel/ssh_tunnel.py
@@ -6,7 +6,7 @@ from typing import Optional, Tuple
 
 from swarmauri_base.transports.peer_mixin import PeerTransportMixin
 from swarmauri_base.transports.runnable_mixin import RunnableMixin
-from swarmauri_base.transports.transport_base import TransportBase
+from swarmauri_base.transports.TransportBase import TransportBase
 from swarmauri_base.transports.unicast_mixin import UnicastTransportMixin
 from swarmauri_core.transports.capabilities import TransportCapabilities
 from swarmauri_core.transports.enums import (
