+++ pkgs/standards/swarmauri_transport_wsjsonrpcmux/swarmauri_transport_wsjsonrpcmux/ws_jsonrpc_mux.py
@@ -6,7 +6,7 @@ from typing import Any, Awaitable, Callable, Optional
 
 from swarmauri_base.transports.peer_mixin import PeerTransportMixin
 from swarmauri_base.transports.runnable_mixin import RunnableMixin
-from swarmauri_base.transports.transport_base import TransportBase
+from swarmauri_base.transports.TransportBase import TransportBase
 from swarmauri_core.transports.capabilities import TransportCapabilities
 from swarmauri_core.transports.enums import (
     AddressScheme,
